# ARROW-6417: [C++][Parquet] Miscellaneous optimizations yielding slightly better Parquet binary read performance