@wesm 2019-09-04
@pitrou @bkietz I changed the BufferBuilder growth strategy from 1.5x to 2x in this patch. I can break that out into a separate patch and run benchmarks if you'd like -- it's a bit weird to make that kind of change in passing here.