@pitrou 2019-09-04
Yes, better use a separate PR. When I tried locally the benchmark numbers were surprisingly unconvincing.