@wesm 2019-09-05
+1. I'm leaving the JIRA open because the performance issue is not resolved